requirements.txt
asgiref==3.6.0
Django==4.1.6
factory-boy==3.2.1
Faker==16.7.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.1
six==1.16.0
sqlparse==0.4.3
